+++ plugins/commands/plugin/command/uninstall.rb
@@ -8,7 +8,7 @@ module VagrantPlugins
       class Uninstall < Base
         def execute
           opts = OptionParser.new do |o|
-            o.banner = "Usage: vagrant plugin uninstall <name> [-h]"
+            o.banner = "Usage: vagrant plugin uninstall <name> [<name2> <name3> ...] [-h]"
           end
 
           # Parse the options
@@ -17,7 +17,9 @@ module VagrantPlugins
           raise Vagrant::Errors::CLIInvalidUsage, :help => opts.help.chomp if argv.length < 1
 
           # Uninstall the gems
-          argv.each{ |gem| action(Action.action_uninstall, :plugin_name => gem) }
+          argv.each do |gem|
+            action(Action.action_uninstall, :plugin_name => gem)
+          end
 
           # Success, exit status 0
           0